Refactored Code so that now all Acces to the FIXTURE_LIST happens inside fixture.rs, Changed name of artnet_loop to dmx_output_loop
