+++ Interface/src/interfaces.rs
@@ -2,8 +2,7 @@ use std::net::UdpSocket;
 use std::io;
 use std::thread::sleep;
 use std::time::{Duration, Instant};
-use Common::fixture;
-use Common::fixture::{universe_count, ChannelError, MAX_CHANNEL};
+use Common::fixture::{MAX_CHANNEL, calculate_dmx_values};
 use crate::artnet::ArtnetInterface;
 
 const TARGET: &str = "255.255.255.255:6454";
@@ -13,11 +12,11 @@ pub trait DmxInterface {
     fn send_universe(&self, local_universe_index: u16, data: &[u8;MAX_CHANNEL as usize]) -> Result<(), io::Error>;
 }
 
-pub fn artnet_loop() -> io::Result<()> {
+pub fn dmx_output_loop() -> io::Result<()> {
     let socket = UdpSocket::bind("0.0.0.0:0")?;
     socket.set_broadcast(true)?;
     
-    let artnet_interface = ArtnetInterface::new(socket, TARGET.to_string());
+    let artnet_interface: Box<dyn DmxInterface> = Box::new(ArtnetInterface::new(socket, TARGET.to_string()));
 
     println!("Starting artnet");
 
@@ -38,37 +37,4 @@ pub fn artnet_loop() -> io::Result<()> {
 
 
     }
-}
-
-pub fn calculate_dmx_values() -> Vec<[u8;MAX_CHANNEL as usize]>{
-    let universe_count = universe_count();
-
-    let mut output = vec![[0u8; MAX_CHANNEL as usize]; universe_count];
-
-    let list = fixture::FIXTURE_LIST.read().unwrap();
-
-    list.fixtures.iter().for_each(|(_, fixture)| {
-        let universe_number = fixture.get_universe();
-        let fixture_type = fixture.get_fixture_type();
-        let fixture_name = fixture.get_name();
-
-        if universe_number < universe_count {
-            fixture
-                .get_channel_values()
-                .iter()
-                .for_each(|(channel, value)| {
-                    *output.get_mut(universe_number).unwrap()
-                        .get_mut(*channel as usize)
-                            .ok_or(ChannelError::ChannelOutOfRange).expect(
-                            &format!("Fixture \"{}\" of type {} has a channel that is out of bounds",
-                                     fixture_name, fixture_type
-                            ))
-                            = *value;
-
-                });
-        }
-
-    });
-
-    output
 }
